backend/app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
babel==2.17.0
beautifulsoup4==4.13.4
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
courlan==1.3.2
cssselect==1.3.0
dateparser==1.2.2
fastapi==0.115.14
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
htmldate==1.9.3
huggingface-hub==0.33.2
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.1
jusText==3.0.2
keybert==0.9.0
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
readability-lxml==0.8.4.1
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
tokenizers==0.21.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.53.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0